ClanEventPersistentForUser.java - user monster ids are longs and made convenience method to return them Clan.proto - PersistentClanEventUserInfoProto has repeated FullUserMonsterProto userMonsters not repeated ints EventClan.proto - BeginClanRaidRequestProto sends repeated FullUserMonsterProto instead of repeated int32 for userMonsters ClanEventPersistentForUserRetrieveUtils.java - user monster ids are longs MonsterForUserRetrieveUtils.java - method to getSpecificUserMonsters(List<Long> userMonsterIds) BeginClanRaidController.java - userMonster ids are longs StartupController.java - sending user monsters of clan users who are in a raid CreateInfoProtoUtils.java - PersistentClanEventUserInfoProto now have repeated FullUserMonsterProto userMonsters instead of repeated ints InsertUtil.java - user monster ids are long

+++ src/main/java/com/lvl6/retrieveutils/ClanEventPersistentForUserRetrieveUtils.java
@@ -84,9 +84,9 @@ import com.lvl6.utils.DBConnection;
 		int crsDmgDone = rs.getInt(i++);
 		int crsmId = rs.getInt(i++);
 		int crsmDmgDone = rs.getInt(i++);
-		int userMonsterIdOne = rs.getInt(i++);
-		int userMonsterIdTwo = rs.getInt(i++);
-		int userMonsterIdThree = rs.getInt(i++);
+		long userMonsterIdOne = rs.getLong(i++);
+		long userMonsterIdTwo = rs.getLong(i++);
+		long userMonsterIdThree = rs.getLong(i++);
 		
 		return new ClanEventPersistentForUser(userId, clanId, crId, crDmgDone, crsId,
 				crsDmgDone, crsmId, crsmDmgDone, userMonsterIdOne, userMonsterIdTwo,
